refactor: extract telegram dir browser.
